# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: ProductivityPal API
  description: AI-помощник для умного общения и управления задачами
  version: 1.0.0
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: "Регистрация пользователя"
      description: "Создание нового аккаунта с выбором режима приватности"
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - username
                - privacy_mode
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                username:
                  type: string
                privacy_mode:
                  type: string
                  enum: [standard, private]
      responses:
        '201':
          description: "Пользователь успешно зарегистрирован"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  access_token:
                    type: string
        '400':
          description: "Неверные данные"
  /auth/login:
    post:
      summary: "Вход в систему"
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: "Успешный вход"
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    default: "bearer"
                  user_id:
                    type: string
        '401':
          description: "Неверные учетные данные"

  # Управление пользователями
...